requirements.txt

Flask
pandas
numpy
matplotlib
seaborn
mysql-connector-python

